Rename remaining Enum examples to Varian examples.
